move tslibs-specific tests from test_timezones
